requirements.txt
aiohttp==3.7.4.post0
arrow==1.1.1
async-timeout==3.0.1
attrs==21.2.0
binaryornot==0.4.4
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.3
fastapi==0.63.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==3.2
Jinja2==3.0.1
jinja2-time==0.2.0
loguru==0.5.3
MarkupSafe==2.0.1
multidict==5.1.0
nb-cli==0.4.2
nonebot-adapter-cqhttp==2.0.0a13
nonebot2==2.0.0a13.post1
Pillow==9.1.1
poyo==0.5.0
prompt-toolkit==1.0.14
pydantic==1.8.2
pyfiglet==0.8.post1
Pygments==2.9.0
pygtrie==2.4.2
PyInquirer==1.0.3
python-dateutil==2.8.2
python-dotenv==0.19.0
python-slugify==5.0.2
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.13.6
text-unidecode==1.3
tomlkit==0.7.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.13.4
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
win32-setctime==1.0.3
yarl==1.6.3
peewee==3.15.1
ruamel-yaml==0.17.21
